+++ src/test/java/com/geophile/z/index/TestSpatialObject.java
@@ -17,7 +17,7 @@ public class TestSpatialObject implements SpatialObject
     }
 
     @Override
-    public long[] arbitraryPoint()
+    public double[] arbitraryPoint()
     {
         fail();
         return null;
